MRN-871 Maked changes to the payment service change the storage of money from integers to hundredths

See merge request Andersen/meetingroom-new/meetingroom-new-backend!157

+++ payment-service/src/main/java/com/andersen/banking/meeting_api/dto/AccountRegistrationDto.java
@@ -49,5 +49,5 @@ public class AccountRegistrationDto {
     @NotNull
     @Schema(example = EXAMPLE_BALANCE, defaultValue = EXAMPLE_BALANCE)
     @JsonProperty("balance")
-    private double balance;
+    private Long balance;
 }
